:wrench: rename guildConfig file for clarity
