+++ src/database/fagc/guildconfig.ts
